Validação da entity pessoa
